requirements.txt
Django==4.2.7
mysqlclient==2.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
django-channels==0.7.0
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
redis==5.0.1
